Added "B" Spells and Martial Classes
I think that, rather than have "-" as the Max Purchases, we should remove that property, like we do with non-material spells and abilities. Also, Blood & Thunder doesn't have a material component, but the white strip is required, so I put it in. All class names are capitalized, and Charge counts have Xs instead of asterisks, as well.